Move tests and stories into __tests__ and __stories__ directories
