+++ typescript/common-resolvers/src/utils/__tests__/get-slug.test.ts
@@ -1,4 +1,4 @@
-import { getSlug } from "./get-slug";
+import { getSlug } from "../get-slug";
 
 type TestCase = [string, string];
